refactor: 도메인 별로 쿼리 훅 구조 분리 (#30)
