+++ src/hooks/queries/useAdminClubQueries.ts
@@ -1,64 +1,8 @@
 import { apiRequest } from '@/api/apiRequest';
-import {
-    IClubRegisterValue,
-    IRecruitNoticeValue,
-    ISummaryInfoValue,
-} from '@/types';
-import convertImageToFile from '@/utils/convertImageToFile';
-import { useQuery } from '@tanstack/react-query';
+import { queryClient } from '@/config/queryClient';
+import { ClubIdType, IRecruitNoticeValue, ISummaryInfoValue } from '@/types';
+import { useMutation, useQuery } from '@tanstack/react-query';
 import { useEffect } from 'react';
-
-// 동아리 등록 정보 불러오기
-const useRegisterInfoQuery = ({
-    clubId,
-    setInputValue,
-    setPreviewImg,
-    setPostImg,
-}: {
-    clubId: number | null;
-    setInputValue: React.Dispatch<React.SetStateAction<IClubRegisterValue>>;
-    setPreviewImg: React.Dispatch<
-        React.SetStateAction<string | ArrayBuffer | null>
-    >;
-    setPostImg: React.Dispatch<React.SetStateAction<File | null>>;
-}) => {
-    const { isSuccess, data, isError } = useQuery({
-        queryKey: [clubId, 'registerInfo'],
-        queryFn: async () => {
-            return await apiRequest({
-                url: `/api/clubs/${clubId}`,
-                method: 'GET',
-            });
-        },
-        staleTime: 5 * 60 * 1000,
-    });
-
-    // 데이터 불러오기 성공 시, 등록 정보 상태 업데이트
-    useEffect(() => {
-        if (isSuccess && data) {
-            setInputValue({
-                clubName: data.result.name,
-                leaderEmail: data.result.leaderEmail,
-                category: data.result.category,
-                oneLiner: data.result.description,
-            });
-
-            console.log(data);
-
-            // 이미지 미리보기 업데이트
-            setPreviewImg(data.result.profileImageUrl);
-
-            // 이미지 파일로 변환 후 상태 업데이트
-            convertImageToFile(data.profileImageUrl).then((imageFile) => {
-                setPostImg(imageFile || null);
-            });
-        }
-
-        if (isError) {
-            console.error('동아리 요약 정보 불러오기 실패');
-        }
-    }, [isSuccess, data]);
-};
 
 // 동아리 요약 정보 불러오기
 const useSummaryInfoQuery = ({
@@ -104,6 +48,33 @@ const useSummaryInfoQuery = ({
     };
 };
 
+// 요약 정보 저장
+const useSaveSummaryInfoMutation = ({
+    clubId,
+    inputValue,
+}: {
+    clubId: ClubIdType;
+    inputValue: ISummaryInfoValue;
+}) =>
+    useMutation({
+        mutationFn: async () => {
+            return await apiRequest({
+                url: `/api/clubs/club-admin/${clubId}`,
+                method: 'POST',
+                data: inputValue,
+                requireToken: true,
+            });
+        },
+        onSuccess: () => {
+            queryClient.invalidateQueries({
+                queryKey: [clubId, 'summaryInfo'],
+            });
+        },
+        onError: (error) => {
+            console.error('요약정보 저장 실패', error);
+        },
+    });
+
 // 모집안내 정보 불러오기
 const useRecruitNoticeQuery = ({
     clubId,
@@ -144,11 +115,39 @@ const useRecruitNoticeQuery = ({
     };
 };
 
+// 모집안내 저장
+const useSaveRecruitNoticeMutation = ({
+    clubId,
+    inputValue,
+}: {
+    clubId: ClubIdType;
+    inputValue: IRecruitNoticeValue;
+}) =>
+    useMutation({
+        mutationFn: async () => {
+            return await apiRequest({
+                url: `/api/clubs/club-admin/${clubId}/recruitment`,
+                method: 'POST',
+                data: inputValue,
+                requireToken: true,
+            });
+        },
+        onSuccess: () => {
+            queryClient.invalidateQueries({
+                queryKey: [clubId, 'recruitNotice'],
+            });
+        },
+        onError: (error) => {
+            console.error('모집안내 저장 실패', error);
+        },
+    });
+
 function useAdminClubQueries() {
     return {
-        useRegisterInfoQuery,
         useSummaryInfoQuery,
+        useSaveSummaryInfoMutation,
         useRecruitNoticeQuery,
+        useSaveRecruitNoticeMutation,
     };
 }
